Reader: Fix black screen bug
Blackscreening of some pages occured because Aidoku was recieving
a zoom box of 0.0 in cases where the image height was greater
than the zoomable view's bounds height.

To fix this error, initialize the zoomable scroll view with
bounds of the UIScreen which gives a proper frame for the image
to be displayed.

Signed-off-by: kingbri <bdashore3@gmail.com>