Kingfisher: Set memory limits
Kingfisher takes up 25% of the system memory by default at a max
value and there is no limit to the amount of images cached in memory.

This is not good since most iPhones have 4GB of memory and Aidoku
has no limits to take up 1GB of the system ram (or 1.5GB on newer
iDevices).

Force Kingfisher to use a max of 300MB of memory and store 150 images
in memory at a time. The rest of the images are cached to disk.

Signed-off-by: kingbri <bdashore3@gmail.com>